% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{calculate_similarity}
\alias{calculate_similarity}
\title{The title of -calculate similarity-}
\usage{
calculate_similarity(individual1, individual2, method)
}
\arguments{
\item{individual1}{A Numeric Vector.}

\item{individual2}{A Numeric Vector.}

\item{method}{A string.}
}
\value{
A Numeric scalar \code{doipkg_similarity}:
\item{individual1}{Numeric Vector.}
\item{individual2}{Numeric Vector.}
\item{individual1individual2}{Numeric scalar. the sum of \code{individual1} and \code{individual2}}
}
\description{
Here is a brief description
}
\details{
Computes the similarity of \code{individual1} and \code{individual2}.
}
