% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_adopters.R
\name{detect_adopters}
\alias{detect_adopters}
\title{The title of -detect_adopters-}
\usage{
detect_adopters(
  num_agents,
  adj_matrix,
  final_matrix,
  original_data,
  stages,
  stages_name,
  ps_theory,
  adoption_efficacy,
  non_adoption_efficacy,
  method,
  p_prior
)
}
\arguments{
\item{num_agents}{A Numeric scala.}

\item{adj_matrix}{A Numeric Matrix.}

\item{final_matrix}{A Numeric Matrix.}

\item{original_data}{A Numeric Matrix.}

\item{stages}{A list.}

\item{stages_name}{A list.}

\item{ps_theory}{A list.}

\item{adoption_efficacy}{A list.}

\item{non_adoption_efficacy}{A Numeric scalar.}

\item{method}{A str.}

\item{p_prior}{A Numeric scalar.}

\item{result}{A list.}
}
\description{
Here is a brief description
}
