% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_network_matrices}
\alias{generate_network_matrices}
\title{The title of -generate network matrices-}
\usage{
generate_network_matrices(num_agents)
}
\arguments{
\item{num_agents}{A numeric scalar indicating the number of agents to generate.}
}
\value{
A list :

A list containing:
\item{original_data}{A data frame of agent demographic data.}
\item{adjacency_matrix}{A symmetric adjacency matrix indicating random connections between agents.}
\item{similarity_matrix}{A standardized similarity matrix based on Euclidean distance.}
\item{final_matrix}{A weighted matrix combining the adjacency and similarity matrices.}
}
\description{
Here is a brief description
}
