% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innovators.R
\name{get_top_innovators}
\alias{get_top_innovators}
\title{The title of -get innovators-}
\usage{
get_top_innovators(final_matrix, p_innovators, method)
}
\arguments{
\item{final_matrix}{A Numeric Matrix.}

\item{p_innovators}{A Numeric scalar.}

\item{method}{A string.}
}
\value{
A Numeric Vector \code{doipkg_innovators}:
\item{final_matrix}{Numeric Matrix.}
\item{p_innovators}{Numeric Vector.}
\item{final_matrixp_innovators}{Numeric scalar. the calculation of  \code{final_matrix} and \code{p_innovators}}
}
\description{
Here is a brief description
}
\details{
Get the innovators of \code{final_matrix} and \code{p_innovators}.
}
