% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adoption_effectiveness.R
\name{simulate_adoption}
\alias{simulate_adoption}
\title{The title of -adoption effectiveness-}
\usage{
simulate_adoption(
  final_matrix,
  result,
  stages,
  ps_theory,
  p_prior,
  effectiveness,
  non_adoption_efficacy,
  original_data
)
}
\arguments{
\item{final_matrix}{A Numeric Matrix.}

\item{result}{A list.}

\item{stages}{A list.}

\item{ps_theory}{A list.}

\item{p_prior}{A Numeric scalar.}

\item{effectiveness}{A list.}

\item{non_adoption_efficacy}{A Numeric scalar.}

\item{original_data}{A Numeric Matrix.}
}
\description{
Here is a brief description
}
