% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innovators.R
\name{determine_adopter}
\alias{determine_adopter}
\title{Get adopters}
\usage{
determine_adopter(
  total_n,
  adj_matrix,
  final_matrix,
  indices,
  p_innovators,
  approach
)
}
\arguments{
\item{total_n}{Numeric scalar. Total number of nodes in the network.}

\item{adj_matrix}{Numeric matrix. The adjacency matrix representing the network.}

\item{final_matrix}{Numeric matrix. An additional matrix related to the network (optional for this function).}

\item{indices}{Numeric vector. Indices of the nodes to consider for adoption.}

\item{p_innovators}{Numeric scalar. Proportion of total nodes to be classified as innovators.}

\item{approach}{String. The centrality method to use: "counts", "betweeness", or "closeness".}
}
\value{
A list containing:
\describe{
\item{indices}{Numeric vector. Indices of the top adopters.}
\item{values}{Numeric vector. Centrality values of the top adopters.}
\item{approach}{String. The centrality measure used.}
}
}
\description{
This function determines adopters in a network using one of three centrality measures: counts, betweenness, or closeness.
}
\details{
Determines the top innovators in the network based on the specified centrality measure.
}
